requirements.txt
streamlit>=1.29.0
pandas>=2.2.0
gspread>=5.12.0
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
plotly>=5.17.0
numpy>=1.26.0
openpyxl>=3.1.2 